backend/ai_detector/requirements.txt
Django==4.2.7
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
mysqlclient==2.2.0
django-cors-headers==4.3.1
python-decouple==3.8
matplotlib==3.7.2
joblib==1.3.2
ultralytics==8.3.189
imageio==2.31.6
imageio-ffmpeg==0.6.0
pillow-avif-plugin==1.5.2
pillow-heif==1.1.0
